update: rename bottomsheetmenu
