+++ lib/view/create_new_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:mobile/routes.dart';
 import 'package:mobile/view/components/common/button_common.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:mobile/view/components/option_menu.dart';
+import 'package:mobile/view/components/bottom_sheet_menu.dart';
 
 class CreateNewScreen extends StatelessWidget {
   final ImagePicker picker = ImagePicker();
